api/requirements.txt
fastapi
uvicorn 
nltk 
scikit-learn 
python-docx 
PyPDF2
google-genai
python-dotenv
sentence-transformers
python-multipart
presidio-analyzer
presidio-anonymizer
spacy

